Add missing dev dependency "documentation"
The `docs` script relies on `documentation`, however this dependency is
not listed in `devDependencies`.

As a result, running `yarn docs` will fail with

>  `documentation: not found`

Since all other development tools like `jest` and `eslint` are listed as
dependencies, this commit adds `documentation` to the list. This way
the repo stays self-sufficient and all scripts are usable out-of-the-box.